Changed graph to PortfolioGraph
