+++ client/src/styles/css/Graph.css
@@ -1,0 +1,5 @@
+graph-card {
+    width: 90%;
+    height: 75%; /* maintains 4:3 aspect ratio*/
+    /* max-height: 400px; */
+}
