+++ client/src/styles/js/graphData.js
